+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.2.36 - not yet released
 * New features:
 * Fixed bugs:
-  * Main README.md file stills mentioning the web version of Mapiah, which was removed in 0.2.31. [reported by CaverBruce at issue #13]
+  * Main README.md file stills mentioning the web version of Mapiah, which was removed in 0.2.31 [reported by CaverBruce at issue [#13](https://github.com/rsevero/mapiah/issues/13)]
+  * Slash produced by keyboard combinations (for example, AltRght + Q) is not processed as a shortcut to add line segments between selected line segments.
 * Infrastructure maintenance:
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release